Se agrega flujo sin validaciones con router.push()
