+++ pages/dashboard/analisisutilidad.js
@@ -1,21 +1,23 @@
 import { Col, Row} from 'react-bootstrap';
 import Image from 'next/image';
+import {useRouter} from 'next/router';
 
-import Layout from '../components/utilities/Layout';
-import LayoutPage from '../components/utilities/layout-page/LayoutPages';
+import Layout from '../../components/utilities/Layout';
+import LayoutPage from '../../components/utilities/layout-page/LayoutPages';
 
-import StepperISR from '../components/utilities/Steppers/StepperISR';
-import TableResumenISR from '../components/utilities/Tables/TableResumenISR';
-import ButtonPage from '../components/utilities/ButtonPage';
-import GraphicBar from '../components/graphics/Graphic-Bar';
-import GraphicCircle from '../components/graphics/Graphic-Circle';
+import StepperISR from '../../components/utilities/Steppers/StepperISR';
+import TableResumenISR from '../../components/utilities/Tables/TableResumenISR';
+import ButtonPage from '../../components/utilities/ButtonPage';
+import GraphicBar from '../../components/graphics/Graphic-Bar';
+import GraphicCircle from '../../components/graphics/Graphic-Circle';
 //import InputMoney from '../components/utilities/Input/Input-money-pages';
 //import LottieISRCobranza from '../components/Lotties/Lottie-isr-cobranza';
 
-import next from '../public/assets/icons/Next.svg';
-import prev from '../public/assets/icons/Previus.svg';
+import next from '../../public/assets/icons/Next.svg';
+import prev from '../../public/assets/icons/Previus.svg';
 
 export default function AnalisisUtilidad() {
+    const router = useRouter();
     return (
         <Layout>
