+++ pages/dashboard/isrcobranza.js
@@ -1,21 +1,23 @@
 import { Col, Row, Button } from 'react-bootstrap';
 import Image from 'next/image';
+import {useRouter} from 'next/router';
 
-import Layout from '../components/utilities/Layout';
-import LayoutPage from '../components/utilities/layout-page/LayoutPages';
+import Layout from '../../components/utilities/Layout';
+import LayoutPage from '../../components/utilities/layout-page/LayoutPages';
 
-import StepperISR from '../components/utilities/Steppers/StepperISR';
-import InputMoney from '../components/utilities/Input/Input-money-pages';
-import LottieISRDeducible from '../components/Lotties/Lottie-isr-deducible';
+import StepperISR from '../../components/utilities/Steppers/StepperISR';
+import InputMoney from '../../components/utilities/Input/Input-money-pages';
+import LottieISRCobranza from '../../components/Lotties/Lottie-isr-cobranza';
 
-import next from '../public/assets/icons/Next.svg';
-import prev from '../public/assets/icons/Previus.svg';
+import next from '../../public/assets/icons/Next.svg';
+import prev from '../../public/assets/icons/Previus.svg';
 
-export default function ISRDeducible() {
+export default function ISRCobranza() {
+    const router = useRouter();
     return (
         <Layout>
             <Col sm={12} md={12}>
-                <StepperISR num='1' />
+                <StepperISR num='0' />
             </Col>
 
             <LayoutPage>
@@ -23,7 +25,7 @@ export default function ISRDeducible() {
                 <Col sm={12} md={12}>
                     <h1 className="title-card">ISR
                         <span className="span-title-card">
-                            Deducible
+                            Cobranza
                         </span>
                     </h1>
                 </Col>
@@ -34,23 +36,22 @@ export default function ISRDeducible() {
                             <Col sm={12} md={5}>
                                 <div className="div-image-card">
                                     <div className="div-container-lottie-card">
-                                        <LottieISRDeducible id="lottieISRDecucilbe" className="lottie-pages" alt="ISR Deducible" />
+                                        <LottieISRCobranza id="lottieISRCobranza" className="lottie-pages" alt="Unete a tuContapp" />
                                     </div>
                                 </div>
                             </Col>
                             <Col sm={12} md={7}>
                                 <div className="div-container-text-card">
-                                    <p className="p-text-card">El ISR deducible se dará de las deducciones del mes,
-                                                                estas son todo gasto efectivamente pagado relacionado con tu actividad, estos se deberan considerar sin IVA o alguna retención.</p>
-                                    
+                                    <p className="p-text-card">Por el momento solo se puede calcular ejercicios fiscales de año 2022.</p>
+                                    <p className="p-text-card">Se iniciará con los impuestos relacionados al ISR y posteriormente con el IVA.</p>
                                     <div className="div-container-input-card">
-                                        <InputMoney nombre="ISR Deducible" idInput="Input-isrDeducible" />
+                                        <InputMoney nombre="ISR Cobranza" idInput="Input-isrCobranza" />
                                     </div>
                                     <div className="div-container-buttons-card">
-                                        <Button className="btn-pages-np">
+                                        <Button className="btn-pages-np" onClick={()=> router.push('/dashboard/dashboardHome')}>
                                             <Image className="icon-btn-pages" src={prev} alt="Atrás" />
                                         </Button>
-                                        <Button className="btn-pages-np">
+                                        <Button className="btn-pages-np" onClick={()=> router.push('/dashboard/isrdeducible')}>
                                             <Image className="icon-btn-pages" src={next} alt="Siguiente" />
                                         </Button>
