+++ styles/components/_button-pages.scss
@@ -10,6 +10,7 @@ padding: 10px;
 padding-left: 12px;
 padding-right: 12px;
  }
+
 .btn-primary{
     width: auto;
     min-width: 100px;
